add basic information as a tab
